requirements.txt
Django==5.2.7
gunicorn==21.2.0
psycopg[binary]==3.2.12
python-dotenv==1.0.0
pytz==2024.1
Pillow==10.4.0
whitenoise==6.7.0
posthog==3.5.0
django-ratelimit==4.1.0
coverage==7.5.1
bleach==6.1.0
sentry-sdk==2.19.0
